requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
apscheduler
email-validator
aiofiles
requests